+++ config/EURUSD_M1.h
@@ -25,7 +25,7 @@ struct Stg_Force_Params_M1 : StgParams {
     signal_close_method = 0;
     signal_close_level = (float)25.0;
     price_stop_method = 0;
-    price_stop_level = 25.0;
+    price_stop_level = (float)25.0;
     tick_filter_method = 1;
     max_spread = 0;
   }
